+++ api/src/tests/test_src/test_route/testLogRoute.py
@@ -18,7 +18,7 @@ class TestLogRoute(TestSuite):
         Test performing an HTTP GET request on the '/v2/logs' route. This route is redirected to
         '/v2/logs/' by default.
         """
-        response: Response = self.client.get('/v2/logs')
+        response: Response = self.client.get('/v2/logs', headers={'Authorization': 'Bearer j.w.t'})
         headers = response.headers
         self.assertEqual(response.status_code, 302)
         self.assertIn('/v2/logs/', headers.get('Location'))
@@ -28,7 +28,7 @@ class TestLogRoute(TestSuite):
         Test performing an HTTP POST request on the '/v2/logs' route. This route is redirected to
         '/v2/logs/' by default.
         """
-        response: Response = self.client.post('/v2/logs')
+        response: Response = self.client.post('/v2/logs', headers={'Authorization': 'Bearer j.w.t'})
         headers = response.headers
         self.assertEqual(response.status_code, 307)
         self.assertIn('/v2/logs/', headers.get('Location'))
@@ -39,7 +39,7 @@ class TestLogRoute(TestSuite):
         Test performing an HTTP GET request on the '/v2/logs/' route.  This test proves that the endpoint returns
         a list of logs.
         """
-        response: Response = self.client.get('/v2/logs/')
+        response: Response = self.client.get('/v2/logs/', headers={'Authorization': 'Bearer j.w.t'})
         response_json: dict = response.get_json()
         self.assertEqual(response.status_code, 200)
         self.assertEqual(response_json.get('self'), '/v2/logs')
@@ -50,7 +50,7 @@ class TestLogRoute(TestSuite):
         Test performing an HTTP POST request on the '/v2/logs/' route.  This test proves that calling this endpoint
         with an empty request body results in a 400 error code.
         """
-        response: Response = self.client.post('/v2/logs/')
+        response: Response = self.client.post('/v2/logs/', headers={'Authorization': 'Bearer j.w.t'})
         response_json: dict = response.get_json()
         self.assertEqual(response.status_code, 400)
         self.assertEqual(response_json.get('self'), '/v2/logs')
@@ -78,7 +78,8 @@ class TestLogRoute(TestSuite):
         response: Response = self.client.post(
             '/v2/logs/',
             data=request_body,
-            content_type='application/json'
+            content_type='application/json',
+            headers={'Authorization': 'Bearer j.w.t'}
         )
         response_json: dict = response.get_json()
         self.assertEqual(response.status_code, 400)
@@ -109,7 +110,8 @@ class TestLogRoute(TestSuite):
         response: Response = self.client.post(
             '/v2/logs/',
             data=request_body,
-            content_type='application/json'
+            content_type='application/json',
+            headers={'Authorization': 'Bearer j.w.t'}
         )
         response_json: dict = response.get_json()
         self.assertEqual(response.status_code, 200)
@@ -122,7 +124,7 @@ class TestLogRoute(TestSuite):
         Test performing an HTTP GET request on the '/v2/logs/<log_id>' route.  This test proves that trying to
         retrieve a log with an ID that doesn't exist results in a HTTP 400 error.
         """
-        response: Response = self.client.get('/v2/logs/0')
+        response: Response = self.client.get('/v2/logs/0', headers={'Authorization': 'Bearer j.w.t'})
         response_json: dict = response.get_json()
         self.assertEqual(response.status_code, 400)
         self.assertEqual(response_json.get('self'), '/v2/logs/0')
@@ -135,7 +137,7 @@ class TestLogRoute(TestSuite):
         Test performing an HTTP GET request on the '/v2/logs/<log_id>' route.  This test proves that retrieving
         a log with a valid ID results in the log and a 200 status.
         """
-        response: Response = self.client.get('/v2/logs/1')
+        response: Response = self.client.get('/v2/logs/1', headers={'Authorization': 'Bearer j.w.t'})
         response_json: dict = response.get_json()
         self.assertEqual(response.status_code, 200)
         self.assertEqual(response_json.get('self'), '/v2/logs/1')
@@ -147,7 +149,7 @@ class TestLogRoute(TestSuite):
         Test performing an HTTP PUT request on the '/v2/logs/<log_id>' route.  This test proves that trying to
         update a log that doesn't exist results in a 400 error.
         """
-        response: Response = self.client.put('/v2/logs/0')
+        response: Response = self.client.put('/v2/logs/0', headers={'Authorization': 'Bearer j.w.t'})
         response_json: dict = response.get_json()
         self.assertEqual(response.status_code, 400)
         self.assertEqual(response_json.get('self'), '/v2/logs/0')
@@ -160,7 +162,7 @@ class TestLogRoute(TestSuite):
         Test performing an HTTP PUT request on the '/v2/logs/<log_id>' route.  This test proves that if the
         updated log is the same as the original log, a 400 error is returned.
         """
-        response: Response = self.client.get('/v2/logs/1')
+        response: Response = self.client.get('/v2/logs/1', headers={'Authorization': 'Bearer j.w.t'})
         response_json: dict = response.get_json()
         self.assertEqual(response.status_code, 200)
         self.assertIsNotNone(response_json.get('log'))
@@ -170,7 +172,8 @@ class TestLogRoute(TestSuite):
         response: Response = self.client.put(
             '/v2/logs/1',
             data=request_body,
-            content_type='application/json'
+            content_type='application/json',
+            headers={'Authorization': 'Bearer j.w.t'}
         )
         response_json: dict = response.get_json()
         self.assertEqual(response.status_code, 400)
@@ -211,7 +214,8 @@ class TestLogRoute(TestSuite):
         response: Response = self.client.put(
             '/v2/comments/1',
             data=request_body,
-            content_type='application/json'
+            content_type='application/json',
+            headers={'Authorization': 'Bearer j.w.t'}
         )
         response_json: dict = response.get_json()
         self.assertEqual(response.status_code, 200)
@@ -224,7 +228,7 @@ class TestLogRoute(TestSuite):
         Test performing an HTTP DELETE request on the '/v2/logs/<log_id>' route.  This test proves that the
         endpoint should return a 204 success status, no matter if the log existed or not.
         """
-        response: Response = self.client.delete('/v2/logs/0')
+        response: Response = self.client.delete('/v2/logs/0', headers={'Authorization': 'Bearer j.w.t'})
         self.assertEqual(response.status_code, 204)
 
     def test_log_by_id_soft_delete_route_400_no_existing(self) -> None:
@@ -233,9 +237,9 @@ class TestLogRoute(TestSuite):
         the log doesn't exist, a 400 error is returned.
         """
         # Ensure that the log was already deleted before testing the DELETE endpoint
-        self.client.delete('/v2/logs/0')
+        self.client.delete('/v2/logs/0', headers={'Authorization': 'Bearer j.w.t'})
 
-        response: Response = self.client.delete('/v2/logs/soft/0')
+        response: Response = self.client.delete('/v2/logs/soft/0', headers={'Authorization': 'Bearer j.w.t'})
         response_json: dict = response.get_json()
         self.assertEqual(response.status_code, 400)
         self.assertFalse(response_json.get('deleted'))
@@ -261,11 +265,16 @@ class TestLogRoute(TestSuite):
             "time_created": "2019-11-23 16:00:00",
             "deleted": 'Y'
         })
-        response: Response = self.client.post('/v2/logs/', data=request_body, content_type='application/json')
+        response: Response = self.client.post(
+            '/v2/logs/',
+            data=request_body,
+            content_type='application/json',
+            headers={'Authorization': 'Bearer j.w.t'}
+        )
         response_json: dict = response.get_json()
         log_id = response_json.get('log').get('log_id')
 
-        response: Response = self.client.delete(f'/v2/logs/soft/{log_id}')
+        response: Response = self.client.delete(f'/v2/logs/soft/{log_id}', headers={'Authorization': 'Bearer j.w.t'})
         self.assertEqual(response.status_code, 400)
 
     def test_log_by_id_soft_delete_route_204(self) -> None:
@@ -292,14 +301,19 @@ class TestLogRoute(TestSuite):
             "time_created": "2019-11-23 15:00:00",
             "deleted": None
         })
-        response: Response = self.client.post('/v2/logs/', data=request_body, content_type='application/json')
+        response: Response = self.client.post(
+            '/v2/logs/',
+            data=request_body,
+            content_type='application/json',
+            headers={'Authorization': 'Bearer j.w.t'}
+        )
         response_json: dict = response.get_json()
         log_id = response_json.get('log').get('log_id')
 
-        response = self.client.delete(f'/v2/logs/soft/{log_id}')
+        response = self.client.delete(f'/v2/logs/soft/{log_id}', headers={'Authorization': 'Bearer j.w.t'})
         self.assertEqual(response.status_code, 204)
 
-        response: Response = self.client.get(f'/v2/logs/{log_id}')
+        response: Response = self.client.get(f'/v2/logs/{log_id}', headers={'Authorization': 'Bearer j.w.t'})
         response_json: dict = response.get_json()
         self.assertEqual(response.status_code, 200)
         self.assertIsNotNone(response_json.get('log'))
